Split endpoints into their own file to avoid cyclic imports in the next commit
