+++ cassandra/metadata.py
@@ -40,7 +40,7 @@ from cassandra.protocol import QueryMessage
 from cassandra.query import dict_factory, bind_params
 from cassandra.util import OrderedDict, Version
 from cassandra.pool import HostDistance
-from cassandra.connection import EndPoint
+from cassandra.endpoint import EndPoint
 
 log = logging.getLogger(__name__)
